backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.6.3
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
eventlet==0.35.2
flask-socketio==5.3.6
